+++ src/main/java/api/m2/movements/enums/BadgeType.java
@@ -1,0 +1,5 @@
+package api.m2.movements.enums;
+
+public enum BadgeType {
+    BUDGET_MET
+}